% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{arrumar}
\alias{arrumar}
\title{Arruma o arquivo num BD pronto para predição.}
\usage{
arrumar(arqs_treino, nm)
}
\arguments{
\item{arqs_treino}{arquivo com a imagem em jpeg.}

\item{nm}{nomes do modelo para completar a base de predição.}
}
\description{
Deve ser usada dentro da função predizer.
}

