% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{converter_em_matriz}
\alias{converter_em_df}
\alias{converter_em_matriz}
\title{Converter}
\usage{
converter_em_matriz(img)

converter_em_df(m)
}
\arguments{
\item{img}{imagem em formato de data.frame com as coluans x, y, r,g, e b}
}
\description{
Converter de data.frame p/ matriz e de matriz p/ data.frame.

Converter em data.frame
}

