% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{preencher}
\alias{limpar}
\alias{preencher}
\title{Preencher}
\usage{
preencher(img, x, y, lim, x_min = 0, y_min = 0)

limpar(img, n = 1, k = 6, x = 170, y = 30, lim = 0)
}
\arguments{
\item{img}{imagem em formato de data.frame com as coluans x, y, r,g, e b}

\item{x}{x máximo}

\item{y}{y máximo}

\item{lim}{tom limite de cor [0,1]. quanto mais proximo de 0, mais escuro. Ou seja
quanto menor o valor dessa variável, mais fácil de desconsiderar aquele ponto.
Como o default é 0, significa que apenas são considerados os pontos que são totalmente pretos.}

\item{n}{raio da vizinhança utilizado. 1 utiliza apenas as células adjacentes.}

\item{k}{número de pontos escuros necessários no entorno, para que este seja mantido.}
}
\description{
Preencher

Limpar imagem
}
\note{
se você estiver usando a função \code{limpar_por_posicao}, é importante
que o data.frame contenha uma coluna \code{posicao}
}

