% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predizer}
\alias{predizer}
\title{Predizer os numeros do arquivo}
\usage{
predizer(a)
}
\arguments{
\item{a}{arquivo com a imagem em jpeg.}
}
\value{
character com o número predito.
}
\description{
Recebe um arquivo jpeg e retorna o texto do resultado predito.
}

